requirements.txt
Django
django-allauth
djangorestframework
